requirements.txt
uvicorn[standard]
fastapi
mysql-connector-python
python-dotenv
